Fix issue with tests and utils/utils
