+++ utils/storage_utils.ts
@@ -2,6 +2,9 @@
 // See LICENSE.txt for license information.
 
 import {GlobalState} from 'types/store';
+import {DraftInfo} from 'types/store/draft';
+
+import {StoragePrefixes} from './constants';
 
 export function getPrefix(state: GlobalState) {
     if (state && state.entities && state.entities.users && state.entities.users.profiles) {
@@ -14,3 +17,21 @@ export function getPrefix(state: GlobalState) {
     return 'unknown_';
 }
 
+export function getDraftInfoFromKey(key: string, prefix: string): DraftInfo | null {
+    const keyArr = key.split('_');
+    if (prefix === StoragePrefixes.DRAFT) {
+        return {
+            id: keyArr[1],
+            type: 'channel',
+        };
+    }
+
+    if (prefix === StoragePrefixes.COMMENT_DRAFT) {
+        return {
+            id: keyArr[2],
+            type: 'thread',
+        };
+    }
+
+    return null;
+}
